+++ backend/green-india/src/main/java/com/green_india/entity/User.java
@@ -76,8 +76,8 @@ public class User implements UserDetails {
 
 
     // --- Existing Getters / Setters ---
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
     public String getEmail() { return email; }
